refactor(revm): isolate native multisig validation
